requirements-dev.txt
-e .
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
anywidget==0.9.13
appdirs==1.4.4
appnope==0.1.4 ; platform_system == 'Darwin'
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arro3-compute==0.4.5
arro3-core==0.4.5
arro3-io==0.4.5
arrow==1.3.0
asn1crypto==1.5.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1 ; python_full_version < '3.11'
atpublic==5.0
attrs==24.3.0
babel==2.16.0
beartype==0.19.0
beautifulsoup4==4.12.3
bitarray==3.0.0
black==24.10.0
bleach==6.2.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
clickhouse-connect==0.8.10
cloudpickle==3.1.0
codespell==2.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.9
cryptography==44.0.0
cycler==0.12.1
databricks-sql-connector-core==4.0.1
datafusion==43.1.0
db-dtypes==1.3.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deltalake==0.22.3
distlib==0.3.9
doit==0.36.0
duckdb==1.1.3
dunamai==1.23.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2 ; python_full_version < '3.11'
execnet==2.1.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gcsfs==2024.10.0
geoarrow-types==0.2.0
geopandas==1.0.1
google-api-core==2.24.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1 ; (python_full_version < '3.13' and platform_machine == 'AMD64') or (python_full_version < '3.13' and platform_machine == 'WIN32') or (python_full_version < '3.13' and platform_machine == 'aarch64') or (python_full_version < '3.13' and platform_machine == 'amd64') or (python_full_version < '3.13' and platform_machine == 'ppc64le') or (python_full_version < '3.13' and platform_machine == 'win32') or (python_full_version < '3.13' and platform_machine == 'x86_64')
griffe==0.49.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hypothesis==6.122.3
identify==2.6.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
impyla==0.19.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
itables==2.2.4
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-cache==1.0.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.3
jupyterlab-night==0.5.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
jupyterlite-core==0.3.0
jupyterlite-pyodide-kernel==0.3.2
kiwisolver==1.4.7
lonboard==0.10.3
lz4==4.3.3
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mizani==0.13.1
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.6
narwhals==1.18.4
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook-shim==0.2.4
numpy==2.2.0
oauthlib==3.2.2
openpyxl==3.1.5
oracledb==2.5.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
parsy==2.1
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pillow==11.0.0
pins==0.8.6
pkginfo==1.12.0
platformdirs==4.3.6
plotly==5.24.1
plotnine==0.14.4
pluggy==1.5.0
plum-dispatch==2.5.4
polars==1.17.1
pprintpp==0.4.0
pre-commit==4.0.1
prometheus-client==0.21.1
prompt-toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
psycopg2==2.9.10
psygnal==0.11.1
ptyprocess==0.7.0 ; os_name != 'nt' or (sys_platform != 'emscripten' and sys_platform != 'win32')
pure-eval==0.2.3
pure-sasl==0.6.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-core==2.27.1
pydata-google-auth==1.9.0
pydruid==0.6.9
pyexasol==0.27.0
pygments==2.18.0
pyinstrument==5.0.0
pyjwt==2.10.1
pyodbc==5.2.0
pyogrio==0.10.0
pyopenssl==24.3.0
pyparsing==3.2.0
pyproj==3.7.0
pyspark==3.5.3
pytest==8.3.4
pytest-benchmark==5.1.0
pytest-clarity==1.0.1
pytest-cov==6.0.0
pytest-deadfixtures==2.2.1
pytest-httpserver==1.1.0
pytest-mock==3.14.0
pytest-randomly==3.16.0
pytest-repeat==0.9.3
pytest-snapshot==0.9.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pywin32==308 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
pywinpty==2.0.14 ; os_name == 'nt'
pyyaml==6.0.2
pyzmq==26.2.0
quartodoc==0.9.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.3
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
send2trash==1.8.3
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
soupsieve==2.6
sphobjinv==2.3.1.1
sqlalchemy==2.0.36
sqlglot==26.0.0
stack-data==0.6.3
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
thrift==0.16.0
thrift-sasl==0.4.3
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trino==0.331.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.28.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
